# Remove StringBuilder from HebrewNumber formatting